WIP: applied suggestions from the review
